requirements.txt
# FastAPI Framework and Server
fastapi>=0.100.0
uvicorn[standard]>=0.22.0

# Pydantic for Data Validation and Settings
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0  # Required for Pydantic EmailStr

# HTTP and Async Operations
httpx>=0.24.0  # For async HTTP requests (Bland, HubSpot, n8n, etc.)

# Environment and Configuration
python-dotenv>=1.0.0  # For loading .env files

# Logging
logfire>=0.16.0  # For structured logging

# AI and LLM (Marvin is optional based on LLM_PROVIDER)
marvin>=1.2.0 # If using Marvin as LLM_PROVIDER
# openai>=1.0.0 # If using OpenAI directly
# anthropic>=0.20.0 # If using Anthropic directly
# google-generativeai>=0.4.0 # If using Gemini directly

# Date and Time Handling
python-dateutil>=2.8.2

# System Monitoring (for health check)
psutil>=5.9.0

# Markdown Rendering (for serving docs as HTML)
markdown>=3.4.0
pymdown-extensions>=9.0

# Redis Client (for caching and dashboard data)
redis>=4.6.0

# Google API Libraries
google-api-python-client>=2.80.0  # For Google Sheets API
google-auth-httplib2>=0.1.0       # For Google API authentication
google-auth-oauthlib>=0.7.0        # For Google API authentication
googlemaps>=4.10.0               # For Google Maps Distance Matrix API (Adjusted version)

# Optional: Scheduler (if background tasks need more advanced scheduling)
# fastapi-scheduler>=0.4.0 # Commented out due to version issue

# Optional: For specific utilities if used more broadly
geopy>=2.3.0
pandas>=2.0.0

# Additional Dependencies
python-jose[cryptography]>=3.3.0
bcrypt>=3.2  # Add bcrypt explicitly
motor>=3.0  # Add async MongoDB driver
